+++ bilrost-derive/src/field/value.rs
@@ -270,7 +270,7 @@ impl Field {
         if self.in_oneof {
             vec![
                 quote!(#ty: ::bilrost::encoding::ValueEncoder<#encoder>),
-                quote!(#ty: ::bilrost::encoding::NewForOverwrite),
+                quote!(#ty: ::bilrost::encoding::ForOverwrite),
             ]
         } else {
             vec![
@@ -290,7 +290,7 @@ impl Field {
         if self.in_oneof {
             vec![
                 quote!(#ty: ::bilrost::encoding::DistinguishedValueEncoder<#encoder>),
-                quote!(#ty: ::bilrost::encoding::EmptyState),
+                quote!(#ty: ::bilrost::encoding::ForOverwrite),
                 quote!(#ty: ::core::cmp::Eq),
             ]
         } else {